+++ src/component/page/MainPage.jsx
@@ -1,35 +1,38 @@
-import React, {useState} from 'react';
-import Sidebar from './Sidebar';
-import {Box} from '@mui/material';
-import Header from '../ui/Header';
-import ContentGrid from '../ui/ContentGrid';
+import React from 'react';
+import { useMediaQuery } from 'react-responsive';
+import MainPageContent from '../ui/MainPageContent';
+
 
 function MainPage(props) {
-    
-    const [toggle, setToggle] = useState(false);
 
-    const handleDrawerOpen = () => {
-        setToggle(true);
-    }
-    
-    const handleDrawerClose = () => {
-        setToggle(false);
-    }
+    const isFull = useMediaQuery({
+        query : "(min-width:1440px)"
+    });
+
+    const isLarge = useMediaQuery({
+        query : "(min-width:1250px) and (max-width:1439px)"
+    });
+
+    const isLargeMedium = useMediaQuery({
+        query : "(min-width:850px) and (max-width:1249px)"
+    });
+
+    const isMedium = useMediaQuery({
+        query : "(min-width:650px) and (max-width:849px)"
+    });
+
+    const isSmall = useMediaQuery({
+        query : "(max-width:649px)"
+    });
     
     return (
-        <Box
-            sx={{
-                width: '172%',
-                height: '86vh',
-                border: '1px solid skyblue',
-                padding: '4.5%',
-                paddingBottom: '1%' 
-            }}
-        >
-            <Header setToggle={setToggle} openBar={handleDrawerOpen}/>
-            <Sidebar toggle={toggle} closeBar={handleDrawerClose}/>
-            <ContentGrid></ContentGrid>
-        </Box>
+        <div>
+            {isFull && <MainPageContent style={{ marginTop: '0.2%' }}></MainPageContent>}
+            {isLarge && <MainPageContent style={{ marginTop: '2%' }}></MainPageContent>}
+            {isLargeMedium && <MainPageContent style={{ marginTop: '3.3%' }}></MainPageContent>}
+            {isMedium && <MainPageContent style={{ marginTop: '6%' }}></MainPageContent>}
+            {isSmall && <MainPageContent style={{ marginTop: '8%' }}></MainPageContent>}
+        </div>
     );
 }
